Split session item reads and writes
